+++ caravel/const.py
@@ -1,7 +1,7 @@
 """ Package constants """
 
 import os
-from peppy import COMPUTE_SETTINGS_VARNAME
+from divvy.const import COMPUTE_SETTINGS_VARNAME
 from looper import __version__ as looper_version
 
 
@@ -28,6 +28,6 @@ CONFIG_PRJ_KEY = "projects"
 CONFIG_TOKEN_KEY = "token"
 TOKEN_FILE_NAME = ".caravel_token"
 TOKEN_LEN = 15
-SET_ELSEWHERE = [["--sp"], ["--compute"], ["--env"], ["--help"], ["--version"], ["--selector-attribute"], ["--selector-exclude"],
-                 ["--selector-include"], ['']]
-
+SET_ELSEWHERE = [["--sp"], ["--compute"], ["--env"], ["--help"], ["--version"], ["--selector-attribute"],
+                 ["--selector-exclude"], ["--selector-include"], ['']]
+LOG_FILENAME = "caravel.log"