+++ caravel/helpers.py
@@ -8,7 +8,12 @@ import os
 import random
 import string
 import sys
-from peppy.utils import coll_like
+import looper
+import peppy
+import fcntl
+import termios
+import struct
+from const import *
 
 
 def eprint(*args, **kwargs):
@@ -29,9 +34,11 @@ def expand_path(p, root=""):
     if root:
         if not os.path.isabs(root):
             raise ValueError("Non-absolute root path: {}".format(root))
+        
         def absolutize(x):
             return os.path.join(root, x)
     else:
+        
         def absolutize(x):
             return x
     exp = os.path.expanduser(os.path.expandvars(p))
@@ -64,7 +71,7 @@ def glob_if_exists(x):
     :param x: a string with path containing globs
     :return list[str]: a list of paths
     """
-    return [glob.glob(e) or e for e in x] if coll_like(x) else (glob.glob(x) or [x])
+    return [glob.glob(e) or e for e in x] if peppy.utils.coll_like(x) else (glob.glob(x) or [x])
 
 
 def random_string(n):
@@ -105,7 +112,7 @@ class CaravelParser(argparse.ArgumentParser):
                  "CARAVEL will be used instead.")
 
         self.add_argument(
-            "-d", "--debug-mode",
+            "-d", "--dbg",
             action="store_true",
             dest="debug",
             help="Use this option if you want to enter the debug mode. Unsecured.")
@@ -120,3 +127,74 @@ def _version_text(sep):
     from looper import __version__ as looper_version
     return "caravel version: {}".format(caravel_version) + sep + \
            "looper version: {}\n".format(looper_version)
+
+
+def print_terminal_width(txt=None, char="-"):
+    """
+    Print a line composed of the chars and a text in the middle of the terminal
+
+    :param str txt: a string to display in the middle of the terminal
+    :param str char: a character that the box will be composed of
+    """
+    char = str(char)
+    assert len(char) == 1, "The length of the char parameter has to be equal 1, got '{}'".format(len(char))
+    spaced_txt = txt.center(len(txt)+2) if txt is not None else ""
+    fill_width = int(0.5 * (terminal_width() - len(spaced_txt)))
+    filler = char * fill_width
+    print(filler + spaced_txt + filler)
+
+
+def terminal_width():
+    """
+    Get terminal width
+
+    :return: width of the terminal
+    :rtype int
+    """
+    _, tw = struct.unpack('HH', fcntl.ioctl(0, termios.TIOCGWINSZ, struct.pack('HH', 0, 0)))
+    return tw
+
+
+def run_looper(prj, args, act, log_path, logging_lvl):
+    """
+    Prepare and run looper action using the provided arguments
+
+    :param looper.project.Project prj: project to execute looper action on
+    :param argparse.Namespace args: set of looper arguments
+    :param str act: action to run
+    :param str log_path: absolute path to the log file location
+    :param int logging_lvl: logging level code
+    """
+    # Establish looper logger
+    looper.setup_looper_logger(level=logging_lvl, additional_locations=log_path)
+    # run selected looper action
+    with peppy.ProjectContext(prj) as prj:
+        if act == "run":
+            run = looper.looper.Runner(prj)
+            try:
+                print_terminal_width("looper log")
+                run(args, None)
+                print_terminal_width()
+            except IOError:
+                raise Exception("{} pipelines_dir: '{}'".format(
+                    prj.__class__.__name__, prj.metadata.pipelines_dir))
+
+        if act == "destroy":
+            print_terminal_width("looper log")
+            looper.looper.Destroyer(prj)(args)
+            print_terminal_width()
+
+        if act == "summarize":
+            print_terminal_width("looper log")
+            looper.looper.Summarizer(prj)()
+            print_terminal_width()
+
+        if act == "check":
+            print_terminal_width("looper log")
+            looper.looper.Checker(prj)(flags=args.flags)
+            print_terminal_width()
+
+        if act == "clean":
+            print_terminal_width("looper log")
+            looper.looper.Cleaner(prj)(args)
+            print_terminal_width()