+++ docs/changelog.md
@@ -5,11 +5,21 @@
 Changed: 
 
   - Use logger instead of printing to the error for logging
+  - Instead of having the optional `token` section in the config file 
+  allow to use hidden `.caravel_token` to predefine the token 
 
 Added:
 
   - New `bootstrap` UI
+  - Back button to each page
   - Compute environment configuration page
+  - Deactivate subproject with a dedicated button
+  - Use `divvy` for computing environment configuration
+  - Display `looper` log in the terminal and in a separate page, when finished
+  - `looper` version assurance
+  - New way of `caravel`, `looper` and `python` versions monitoring on the client side in the footer
+  - Reactive button for summary page display (produced by `looper summarize`)
+  - Systematic HTML form elements determination based on `looper` arguments options 
 
 ## v0.1.1 (*2019-01-15*)